// quizonline-frontend/src/app/api/generated/model/plan-enum.ts
/**
 * QuizOnline API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * * `free` - Free * `paid` - Paid
 */
export enum PlanEnumDto {

    Free = 'free',

    Paid = 'paid'
}